requirements.txt
streamlit==1.29.0
pandas<2.1.0
numpy==1.26.2
plotly==5.18.0
pillow==10.1.0
scikit-learn==1.3.2
python-dotenv==1.0.0
watchdog==3.0.0
